+++ indispensable-cosef/src/commonMain/kotlin/at/asitplus/signum/indispensable/cosef/CoseMac.kt
@@ -2,12 +2,8 @@ package at.asitplus.signum.indispensable.cosef
 
 import at.asitplus.KmmResult
 import at.asitplus.catching
-import at.asitplus.signum.indispensable.CryptoSignature
 import at.asitplus.signum.indispensable.contentEqualsIfArray
-import at.asitplus.signum.indispensable.contentHashCodeIfArray
 import at.asitplus.signum.indispensable.cosef.CoseSigned.Companion.create
-import at.asitplus.signum.indispensable.cosef.io.ByteStringWrapper
-import at.asitplus.signum.indispensable.cosef.io.ByteStringWrapperSerializer
 import at.asitplus.signum.indispensable.cosef.io.coseCompliantSerializer
 import io.matthewnelson.encoding.base16.Base16
 import io.matthewnelson.encoding.core.Encoder.Companion.encodeToByteArray
@@ -17,44 +13,35 @@ import kotlinx.serialization.Serializable
 import kotlinx.serialization.encodeToByteArray
 
 /**
- * Representation of a signed COSE_Sign1 object, i.e. consisting of protected header, unprotected header and payload.
- *
- * If the payload is a generic [ByteArray], then it will be serialized as-is. Should the payload be any other type,
- * we will tag it with 24 (see
- * [RFC8949 3.4.5.1](https://www.rfc-editor.org/rfc/rfc8949.html#name-encoded-cbor-data-item)) for serialization.
- * In order to prevent nested wrapping of the payload and the resulting type erasure
- * payloads of type [ByteStringWrapper] will be rejected.
- * In this case the payload could be handed over as the wrapped class itself or manually serialized to [ByteArray]
- *
- * See [RFC 9052](https://www.rfc-editor.org/rfc/rfc9052.html).
+ * Representation of a COSE_MAC0 object with the tag value, i.e. consisting of protected header, unprotected header and payload.
  */
 @OptIn(ExperimentalSerializationApi::class)
-@Serializable(with = CoseSignedSerializer::class)
+@Serializable(with = CoseMacSerializer::class)
 @ConsistentCopyVisibility
-data class CoseSigned<P : Any?> internal constructor(
+data class CoseMac<P : Any?> internal constructor(
     val protectedHeader: CoseHeader,
     val unprotectedHeader: CoseHeader? = null,
     val payload: P?,
-    val signature: CryptoSignature.RawByteEncodable,
+    val tag: ByteArray,
     val wireFormat: CoseBytes,
 ) {
 
     /**
      * @param detachedPayload only to be set when [payload] is null, i.e. it is transported externally
      */
-    fun prepareCoseSignatureInput(
+    fun prepareCoseMacInput(
         externalAad: ByteArray = byteArrayOf(),
         detachedPayload: ByteArray? = null,
-    ): ByteArray = wireFormat.toCoseSignatureInput(externalAad, detachedPayload)
+    ): ByteArray = wireFormat.toCoseMacInput(externalAad, detachedPayload)
 
     fun serialize(parameterSerializer: KSerializer<P>): ByteArray = coseCompliantSerializer
-        .encodeToByteArray(CoseSignedSerializer(parameterSerializer), this)
+        .encodeToByteArray(CoseMacSerializer(parameterSerializer), this)
 
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
         if (other == null || this::class != other::class) return false
 
-        other as CoseSigned<*>
+        other as CoseMac<*>
 
         if (protectedHeader != other.protectedHeader) return false
         if (unprotectedHeader != other.unprotectedHeader) return false
@@ -62,7 +49,7 @@ data class CoseSigned<P : Any?> internal constructor(
             if (other.payload == null) return false
             if (!payload.contentEqualsIfArray(other.payload)) return false
         } else if (other.payload != null) return false
-        if (signature != other.signature) return false
+        if (!tag.contentEquals(other.tag)) return false
         if (wireFormat != other.wireFormat) return false
 
         return true
@@ -71,52 +58,52 @@ data class CoseSigned<P : Any?> internal constructor(
     override fun hashCode(): Int {
         var result = protectedHeader.hashCode()
         result = 31 * result + (unprotectedHeader?.hashCode() ?: 0)
-        result = 31 * result + (payload?.contentHashCodeIfArray() ?: 0)
-        result = 31 * result + signature.hashCode()
+        result = 31 * result + (payload?.hashCode() ?: 0)
+        result = 31 * result + tag.contentHashCode()
         result = 31 * result + wireFormat.hashCode()
         return result
     }
 
     override fun toString(): String {
-        return "CoseSigned(protectedHeader=$protectedHeader," +
+        return "CoseMac(protectedHeader=$protectedHeader," +
                 " unprotectedHeader=$unprotectedHeader," +
                 " payload=${if (payload is ByteArray) payload.encodeToByteArray(Base16()) else payload}," +
-                " signature=$signature," +
+                " tag=$tag," +
                 " wireFormat=$wireFormat)"
     }
 
     companion object {
-        fun <P : Any> deserialize(parameterSerializer: KSerializer<P>, it: ByteArray): KmmResult<CoseSigned<P>> =
+        fun <P : Any> deserialize(parameterSerializer: KSerializer<P>, it: ByteArray): KmmResult<CoseMac<P>> =
             catching {
-                coseCompliantSerializer.decodeFromByteArray(CoseSignedSerializer(parameterSerializer), it)
+                coseCompliantSerializer.decodeFromByteArray(CoseMacSerializer(parameterSerializer), it)
             }
 
         /**
-         * Use this method to create a new [CoseSigned] object with correct [CoseSigned.wireFormat] set.
+         * Use this method to create a new [CoseMac] object with correct [CoseMac.wireFormat] set.
          */
         @Throws(IllegalArgumentException::class)
         fun <P : Any> create(
             protectedHeader: CoseHeader,
             unprotectedHeader: CoseHeader? = null,
             payload: P?,
-            signature: CryptoSignature.RawByteEncodable,
+            tag: ByteArray,
             payloadSerializer: KSerializer<P>,
-        ): CoseSigned<P> = CoseSigned<P>(
+        ): CoseMac<P> = CoseMac<P>(
             protectedHeader = protectedHeader,
             unprotectedHeader = unprotectedHeader,
             payload = payload,
-            signature = signature,
+            tag = tag,
             wireFormat = CoseBytes(
                 protectedHeader = protectedHeader,
                 unprotectedHeader = unprotectedHeader,
                 payload = payload.toRawPayload(payloadSerializer),
-                rawAuthBytes = signature.rawByteArray
+                rawAuthBytes = tag
             ),
         )
 
         /**
-         * Use this method to prepare a [CoseInput] object to calculate the signature,
-         * and then call [create] to create a [CoseSigned] object.
+         * Use this method to prepare a [CoseInput] object to calculate the tag,
+         * and then call [create] to create a [CoseMac] object.
          */
         @Throws(IllegalArgumentException::class)
         fun <P : Any> prepare(
@@ -125,12 +112,10 @@ data class CoseSigned<P : Any?> internal constructor(
             payload: P?,
             payloadSerializer: KSerializer<P>,
         ): CoseInput = CoseInput(
-            contextString = "Signature1",
+            contextString = "MAC0",
             protectedHeader = protectedHeader,
             externalAad = externalAad,
             payload = payload.toRawPayload(payloadSerializer),
         )
     }
 }
-
-