+++ indispensable-cosef/src/commonTest/kotlin/at/asitplus/signum/indispensable/cosef/CoseSerializationTest.kt
@@ -12,6 +12,7 @@ import io.matthewnelson.encoding.base16.Base16
 import io.matthewnelson.encoding.core.Decoder.Companion.decodeToByteArray
 import io.matthewnelson.encoding.core.Encoder.Companion.encodeToString
 import kotlinx.serialization.ExperimentalSerializationApi
+import kotlinx.serialization.SerializationException
 import kotlinx.serialization.builtins.ByteArraySerializer
 import kotlinx.serialization.builtins.serializer
 import kotlinx.serialization.decodeFromByteArray
@@ -35,7 +36,20 @@ class CoseSerializationTest : FreeSpec({
         }
     }
 
-    "Serialization is correct with JSON" {
+    "CoseMac can not be constructed with ByteStringWrapper" {
+        val payload = ByteStringWrapper("StringType")
+        shouldThrow<IllegalArgumentException> {
+            CoseMac.create(
+                protectedHeader = CoseHeader(algorithm = CoseAlgorithm.MAC.HS256),
+                unprotectedHeader = null,
+                payload = payload,
+                tag = byteArrayOf(),
+                payloadSerializer = ByteStringWrapper.serializer(String.serializer())
+            )
+        }
+    }
+
+    "Serialization is correct with JSON CoseSigned" {
         val payload = "This is the content.".encodeToByteArray()
         val cose = CoseSigned.create(
             protectedHeader = CoseHeader(algorithm = CoseAlgorithm.Signature.RS256),
@@ -48,7 +62,20 @@ class CoseSerializationTest : FreeSpec({
         Json.decodeFromString<CoseSigned<ByteArray>>(Json.encodeToString(cose)) shouldBe cose
     }
 
-    "Serialization is correct with JSON for data class" {
+    "Serialization is correct with JSON CoseMac" {
+        val payload = "This is the content.".encodeToByteArray()
+        val cose = CoseMac.create(
+            protectedHeader = CoseHeader(algorithm = CoseAlgorithm.MAC.HS256),
+            unprotectedHeader = null,
+            payload = payload,
+            tag = byteArrayOf(),
+            payloadSerializer = ByteArraySerializer(),
+        )
+
+        Json.decodeFromString<CoseMac<ByteArray>>(Json.encodeToString(cose)) shouldBe cose
+    }
+
+    "Serialization is correct with JSON for data class CoseSigned" {
         val payload = DataClass("This is the content.")
         val cose = CoseSigned.create(
             protectedHeader = CoseHeader(algorithm = CoseAlgorithm.Signature.RS256),
@@ -61,7 +88,20 @@ class CoseSerializationTest : FreeSpec({
         Json.decodeFromString<CoseSigned<DataClass>>(Json.encodeToString(cose)) shouldBe cose
     }
 
-    "Serialization is correct for byte array" {
+    "Serialization is correct with JSON for data class CoseMac" {
+        val payload = DataClass("This is the content.")
+        val cose = CoseMac.create(
+            protectedHeader = CoseHeader(algorithm = CoseAlgorithm.MAC.HS256),
+            unprotectedHeader = null,
+            payload = payload,
+            tag = byteArrayOf(),
+            payloadSerializer = DataClass.serializer(),
+        )
+
+        Json.decodeFromString<CoseMac<DataClass>>(Json.encodeToString(cose)) shouldBe cose
+    }
+
+    "Serialization is correct for byte array CoseSigned" {
         val payload = "This is the content.".encodeToByteArray()
         val cose = CoseSigned.create(
             protectedHeader = CoseHeader(algorithm = CoseAlgorithm.Signature.RS256),
@@ -80,7 +120,27 @@ class CoseSerializationTest : FreeSpec({
         CoseSigned.deserialize(ByteArraySerializer(), serialized).getOrThrow() shouldBe cose
     }
 
-    "Serialization is correct for null" {
+    "Serialization is correct for byte array with CoseMac" {
+        val payload = "This is the content.".encodeToByteArray()
+        val cose = CoseMac.create(
+            protectedHeader = CoseHeader(algorithm = CoseAlgorithm.MAC.HS256),
+            unprotectedHeader = null,
+            payload = payload,
+            tag = byteArrayOf(),
+            payloadSerializer = ByteArraySerializer(),
+        )
+        val serialized = cose.serialize(ByteArraySerializer())
+
+        val serializedString = serialized.encodeToString(Base16Strict).uppercase()
+        serializedString shouldContain "546869732069732074686520636F6E74656E742E" // "This is the content."
+        serializedString shouldContain "8443A10105" // array of 3 bytes that is a map with 4 (the header for HMAC256)
+
+        cose.payload.contentEquals(payload) shouldBe true
+
+        CoseMac.deserialize(ByteArraySerializer(), serialized).getOrThrow() shouldBe cose
+    }
+
+    "Serialization is correct for null CoseSigned" {
         val cose = CoseSigned.create(
             protectedHeader = CoseHeader(algorithm = CoseAlgorithm.Signature.RS256),
             unprotectedHeader = null,
@@ -98,7 +158,25 @@ class CoseSerializationTest : FreeSpec({
         CoseSigned.deserialize(ByteArraySerializer(), serialized).getOrThrow() shouldBe cose
     }
 
-    "Deserialization is correct for byte array" {
+    "Serialization is correct for null CoseMac" {
+        val cose = CoseMac.create(
+            protectedHeader = CoseHeader(algorithm = CoseAlgorithm.MAC.HS256),
+            unprotectedHeader = null,
+            payload = null,
+            tag = byteArrayOf(),
+            payloadSerializer = ByteArraySerializer(),
+        )
+        val serialized = cose.serialize(ByteArraySerializer())
+
+        val serializedString = serialized.encodeToString(Base16Strict).uppercase()
+        serializedString shouldContain "A0F640" // Empty unprotected header; null; begin of tag
+        serializedString shouldContain "8443A10105" // array of 3 bytes that is a map with 4 (the header for HMAC256)
+        cose.payload shouldBe null
+
+        CoseMac.deserialize(ByteArraySerializer(), serialized).getOrThrow() shouldBe cose
+    }
+
+    "Deserialization is correct for byte array CoseSigned" {
         val input = "8445A101390100A054546869732069732074686520636F6E74656E742E43626172"
 
         val cose = CoseSigned.deserialize(ByteArraySerializer(), input.decodeToByteArray(Base16())).getOrThrow()
@@ -106,7 +184,15 @@ class CoseSerializationTest : FreeSpec({
         cose.wireFormat.payload shouldBe "546869732069732074686520636F6E74656E742E".decodeToByteArray(Base16())
     }
 
-    "Deserialization is correct for null payload" {
+    "Deserialization is correct for byte array CoseMac" {
+        val input = "8443A10105A054546869732069732074686520636F6E74656E742E40"
+
+        val cose = CoseMac.deserialize(ByteArraySerializer(), input.decodeToByteArray(Base16())).getOrThrow()
+        cose.payload shouldBe "This is the content.".encodeToByteArray()
+        cose.wireFormat.payload shouldBe "546869732069732074686520636F6E74656E742E".decodeToByteArray(Base16())
+    }
+
+    "Deserialization is correct for null payload CoseSigned" {
         val input = "8445A101390100A0F643626172"
 
         val cose = CoseSigned.deserialize(ByteArraySerializer(), input.decodeToByteArray(Base16())).getOrThrow()
@@ -114,15 +200,31 @@ class CoseSerializationTest : FreeSpec({
         cose.wireFormat.payload shouldBe null
     }
 
-    "Deserialization fails when trying to parse byte array as data class" {
+    "Deserialization is correct for null payload CoseMac" {
+        val input = "8443A10105A0F640"
+
+        val cose = CoseMac.deserialize(ByteArraySerializer(), input.decodeToByteArray(Base16())).getOrThrow()
+        cose.payload shouldBe null
+        cose.wireFormat.payload shouldBe null
+    }
+
+    "Deserialization fails when trying to parse byte array as data class CoseSigned" {
         val input = "8445A101390100A054546869732069732074686520636F6E74656E742E43626172"
 
-        shouldThrow<IllegalStateException> {
+        shouldThrow<SerializationException> {
             CoseSigned.deserialize(DataClass.serializer(), input.decodeToByteArray(Base16())).getOrThrow()
         }
     }
 
-    "Serialization is correct for data class" {
+    "Deserialization fails when trying to parse byte array as data class CoseMac" {
+        val input = "8443A10105A054546869732069732074686520636F6E74656E742E40"
+
+        shouldThrow<SerializationException> {
+            CoseMac.deserialize(DataClass.serializer(), input.decodeToByteArray(Base16())).getOrThrow()
+        }
+    }
+
+    "Serialization is correct for data class CoseSigned" {
         val payload = DataClass("This is the content.")
         val cose = CoseSigned.create(
             protectedHeader = CoseHeader(algorithm = CoseAlgorithm.Signature.RS256),
@@ -141,7 +243,26 @@ class CoseSerializationTest : FreeSpec({
         CoseSigned.deserialize(DataClass.serializer(), serialized).getOrThrow() shouldBe cose
     }
 
-    "Deserialization is correct for data class" {
+    "Serialization is correct for data class CoseMac" {
+        val payload = DataClass("This is the content.")
+        val cose = CoseMac.create(
+            protectedHeader = CoseHeader(algorithm = CoseAlgorithm.MAC.HS256),
+            unprotectedHeader = null,
+            payload = payload,
+            tag = byteArrayOf(),
+            payloadSerializer = DataClass.serializer(),
+        )
+        val serialized = cose.serialize(DataClass.serializer())
+        val serializedString = serialized.encodeToString(Base16Strict).uppercase()
+
+        serializedString shouldContain "546869732069732074686520636F6E74656E742E" // "This is the content."
+        serializedString shouldContain "8443A10105" // array of 3 bytes that is a map with 4 (the header for HMAC256)
+        cose.payload shouldBe payload
+
+        CoseMac.deserialize(DataClass.serializer(), serialized).getOrThrow() shouldBe cose
+    }
+
+    "Deserialization is correct for data class CoseSigned" {
         val input = "8445A101390100A05822D818581EA167636F6E74656E7474546869732069732074686520636F6E74656E742E43626172"
 
         val cose = CoseSigned.deserialize(DataClass.serializer(), input.decodeToByteArray(Base16())).getOrThrow()
@@ -151,10 +272,29 @@ class CoseSerializationTest : FreeSpec({
         // important part is the D818 as tag(24)
     }
 
-    "Deserialization is correct for data class read as byte array" {
+    "Deserialization is correct for data class CoseMac" {
+        val input = "8443A10105A05822D818581EA167636F6E74656E7474546869732069732074686520636F6E74656E742E43626172"
+
+        val cose = CoseMac.deserialize(DataClass.serializer(), input.decodeToByteArray(Base16())).getOrThrow()
+        cose.payload shouldBe DataClass("This is the content.")
+        cose.wireFormat.payload shouldBe "D818581EA167636F6E74656E7474546869732069732074686520636F6E74656E742E"
+            .decodeToByteArray(Base16())
+        // important part is the D818 as tag(24)
+    }
+
+    "Deserialization is correct for data class read as byte array CoseSigned" {
         val input = "8445A101390100A05822D818581EA167636F6E74656E7474546869732069732074686520636F6E74656E742E43626172"
 
         val cose = CoseSigned.deserialize(ByteArraySerializer(), input.decodeToByteArray(Base16())).getOrThrow()
+        cose.wireFormat.payload shouldBe "D818581EA167636F6E74656E7474546869732069732074686520636F6E74656E742E"
+            .decodeToByteArray(Base16())
+        // important part is the D818 as tag(24)
+    }
+
+    "Deserialization is correct for data class read as byte array CoseMac" {
+        val input = "8443A10105A05822D818581EA167636F6E74656E7474546869732069732074686520636F6E74656E742E43626172"
+
+        val cose = CoseMac.deserialize(ByteArraySerializer(), input.decodeToByteArray(Base16())).getOrThrow()
         cose.wireFormat.payload shouldBe "D818581EA167636F6E74656E7474546869732069732074686520636F6E74656E742E"
             .decodeToByteArray(Base16())
         // important part is the D818 as tag(24)
@@ -183,19 +323,19 @@ class CoseSerializationTest : FreeSpec({
         cose.payload shouldBe "This is the content.".encodeToByteArray()
     }
 
-    "CoseSignatureInput is correct for ByteArray" {
+    "CoseInput is correct for ByteArray" {
         val payload = Random.nextBytes(32)
-        val header = CoseHeader(algorithm = CoseAlgorithm.Signature.ES256)
-        val inputManual = coseCompliantSerializer.encodeToByteArray(
-            CoseSignatureInput(
+        var header = CoseHeader(algorithm = CoseAlgorithm.Signature.ES256)
+        val inputManualSignature = coseCompliantSerializer.encodeToByteArray(
+            CoseInput(
                 contextString = "Signature1",
-                protectedHeader = coseCompliantSerializer.encodeToByteArray(header),
+                protectedHeader = header,
                 externalAad = byteArrayOf(),
                 payload = payload
             )
         ).encodeToString(Base16())
 
-        val inputObject = CoseSigned.create(
+        val inputObjectSignature = CoseSigned.create(
             protectedHeader = header,
             payload = payload,
             signature = CryptoSignature.RSA("bar".encodeToByteArray()),
@@ -203,23 +343,44 @@ class CoseSerializationTest : FreeSpec({
         ).prepareCoseSignatureInput(byteArrayOf())
             .encodeToString(Base16())
 
-        inputManual.shouldContain("Signature1".encodeToByteArray().encodeToString(Base16()))
-        inputObject shouldBe inputManual
+        inputManualSignature.shouldContain("Signature1".encodeToByteArray().encodeToString(Base16()))
+        inputObjectSignature shouldBe inputManualSignature
+
+        header = CoseHeader(algorithm = CoseAlgorithm.MAC.HS256)
+        val inputManualMac = coseCompliantSerializer.encodeToByteArray(
+            CoseInput(
+                contextString = "MAC0",
+                protectedHeader = header,
+                externalAad = byteArrayOf(),
+                payload = payload
+            )
+        ).encodeToString(Base16())
+
+        val inputObjectMac = CoseMac.create(
+            protectedHeader = header,
+            payload = payload,
+            tag = byteArrayOf(),
+            payloadSerializer = ByteArraySerializer(),
+        ).prepareCoseMacInput(byteArrayOf())
+            .encodeToString(Base16())
+
+        inputManualMac.shouldContain("MAC0".encodeToByteArray().encodeToString(Base16()))
+        inputObjectMac shouldBe inputManualMac
     }
 
-    "CoseSignatureInput is correct for custom types" {
+    "CoseInput is correct for custom types" {
         val payload = DataClass(Random.nextBytes(32).encodeToString(Base16Strict))
-        val header = CoseHeader(algorithm = CoseAlgorithm.Signature.ES256)
-        val inputManual = coseCompliantSerializer.encodeToByteArray(
-            CoseSignatureInput(
+        var header = CoseHeader(algorithm = CoseAlgorithm.Signature.ES256)
+        val inputManualSignature = coseCompliantSerializer.encodeToByteArray(
+            CoseInput(
                 contextString = "Signature1",
-                protectedHeader = coseCompliantSerializer.encodeToByteArray(header),
+                protectedHeader = header,
                 externalAad = byteArrayOf(),
                 payload = coseCompliantSerializer.encodeToByteArray(ByteStringWrapper(payload)).wrapInCborTag(24),
             )
         ).encodeToString(Base16())
 
-        val inputObject = CoseSigned.create(
+        val inputObjectSignature = CoseSigned.create(
             protectedHeader = header,
             payload = payload,
             signature = CryptoSignature.RSA("bar".encodeToByteArray()),
@@ -227,8 +388,29 @@ class CoseSerializationTest : FreeSpec({
         ).prepareCoseSignatureInput(byteArrayOf())
             .encodeToString(Base16())
 
-        inputManual.shouldContain("Signature1".encodeToByteArray().encodeToString(Base16()))
-        inputObject shouldBe inputManual
+        inputManualSignature.shouldContain("Signature1".encodeToByteArray().encodeToString(Base16()))
+        inputObjectSignature shouldBe inputManualSignature
+
+        header = CoseHeader(algorithm = CoseAlgorithm.MAC.HS256)
+        val inputManualMac = coseCompliantSerializer.encodeToByteArray(
+            CoseInput(
+                contextString = "MAC0",
+                protectedHeader = header,
+                externalAad = byteArrayOf(),
+                payload = coseCompliantSerializer.encodeToByteArray(ByteStringWrapper(payload)).wrapInCborTag(24),
+            )
+        ).encodeToString(Base16())
+
+        val inputObjectMac = CoseMac.create(
+            protectedHeader = header,
+            payload = payload,
+            tag = byteArrayOf(),
+            payloadSerializer = DataClass.serializer(),
+        ).prepareCoseMacInput(byteArrayOf())
+            .encodeToString(Base16())
+
+        inputManualMac.shouldContain("MAC0".encodeToByteArray().encodeToString(Base16()))
+        inputObjectMac shouldBe inputManualMac
     }